+++ HISTORY.rst
@@ -120,6 +120,7 @@ Unreleased
 - #516 Chart Colors Sometimes Using Extended Alternates
 - #386 surface unsupported chartex chart types without dropping
 - feat: #734 add ``<a:snd>`` (sound) support to click-action
+- #487 read/write legacy PowerPoint comments
 
 1.0.2 (2024-08-07)
 ++++++++++++++++++